requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==3.14.0
django-on-heroku==1.1.2
django-redis==5.3.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
exceptiongroup==1.1.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.4.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.5
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2020.1
redis==4.6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.3.1
toml==0.10.1
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
whitenoise==5.2.0
